+++ services/alb/src/main/java/cloud/stackit/sdk/alb/model/UpdateCredentialsResponse.java
@@ -24,10 +24,11 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 
 /** UpdateCredentialsResponse */
 @javax.annotation.Generated(
@@ -60,6 +61,50 @@ public class UpdateCredentialsResponse {
 		this.credential = credential;
 	}
 
+	/**
+	 * A container for additional, undeclared properties. This is a holder for any undeclared
+	 * properties as specified with the 'additionalProperties' keyword in the OAS document.
+	 */
+	private Map<String, Object> additionalProperties;
+
+	/**
+	 * Set the additional (undeclared) property with the specified name and value. If the property
+	 * does not already exist, create it otherwise replace it.
+	 *
+	 * @param key name of the property
+	 * @param value value of the property
+	 * @return the UpdateCredentialsResponse instance itself
+	 */
+	public UpdateCredentialsResponse putAdditionalProperty(String key, Object value) {
+		if (this.additionalProperties == null) {
+			this.additionalProperties = new HashMap<String, Object>();
+		}
+		this.additionalProperties.put(key, value);
+		return this;
+	}
+
+	/**
+	 * Return the additional (undeclared) property.
+	 *
+	 * @return a map of objects
+	 */
+	public Map<String, Object> getAdditionalProperties() {
+		return additionalProperties;
+	}
+
+	/**
+	 * Return the additional (undeclared) property with the specified name.
+	 *
+	 * @param key name of the property
+	 * @return an object
+	 */
+	public Object getAdditionalProperty(String key) {
+		if (this.additionalProperties == null) {
+			return null;
+		}
+		return this.additionalProperties.get(key);
+	}
+
 	@Override
 	public boolean equals(Object o) {
 		if (this == o) {
@@ -69,12 +114,14 @@ public class UpdateCredentialsResponse {
 			return false;
 		}
 		UpdateCredentialsResponse updateCredentialsResponse = (UpdateCredentialsResponse) o;
-		return Objects.equals(this.credential, updateCredentialsResponse.credential);
+		return Objects.equals(this.credential, updateCredentialsResponse.credential)
+				&& Objects.equals(
+						this.additionalProperties, updateCredentialsResponse.additionalProperties);
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(credential);
+		return Objects.hash(credential, additionalProperties);
 	}
 
 	@Override
@@ -82,6 +129,9 @@ public class UpdateCredentialsResponse {
 		StringBuilder sb = new StringBuilder();
 		sb.append("class UpdateCredentialsResponse {\n");
 		sb.append("    credential: ").append(toIndentedString(credential)).append("\n");
+		sb.append("    additionalProperties: ")
+				.append(toIndentedString(additionalProperties))
+				.append("\n");
 		sb.append("}");
 		return sb.toString();
 	}
@@ -124,17 +174,6 @@ public class UpdateCredentialsResponse {
 								UpdateCredentialsResponse.openapiRequiredFields.toString()));
 			}
 		}
-
-		Set<Map.Entry<String, JsonElement>> entries = jsonElement.getAsJsonObject().entrySet();
-		// check to see if the JSON string contains additional fields
-		for (Map.Entry<String, JsonElement> entry : entries) {
-			if (!UpdateCredentialsResponse.openapiFields.contains(entry.getKey())) {
-				throw new IllegalArgumentException(
-						String.format(
-								"The field `%s` in the JSON string is not defined in the `UpdateCredentialsResponse` properties. JSON: %s",
-								entry.getKey(), jsonElement.toString()));
-			}
-		}
 		JsonObject jsonObj = jsonElement.getAsJsonObject();
 		// validate the optional field `credential`
 		if (jsonObj.get("credential") != null && !jsonObj.get("credential").isJsonNull()) {
@@ -160,6 +199,30 @@ public class UpdateCredentialsResponse {
 						public void write(JsonWriter out, UpdateCredentialsResponse value)
 								throws IOException {
 							JsonObject obj = thisAdapter.toJsonTree(value).getAsJsonObject();
+							obj.remove("additionalProperties");
+							// serialize additional properties
+							if (value.getAdditionalProperties() != null) {
+								for (Map.Entry<String, Object> entry :
+										value.getAdditionalProperties().entrySet()) {
+									if (entry.getValue() instanceof String)
+										obj.addProperty(entry.getKey(), (String) entry.getValue());
+									else if (entry.getValue() instanceof Number)
+										obj.addProperty(entry.getKey(), (Number) entry.getValue());
+									else if (entry.getValue() instanceof Boolean)
+										obj.addProperty(entry.getKey(), (Boolean) entry.getValue());
+									else if (entry.getValue() instanceof Character)
+										obj.addProperty(
+												entry.getKey(), (Character) entry.getValue());
+									else {
+										JsonElement jsonElement = gson.toJsonTree(entry.getValue());
+										if (jsonElement.isJsonArray()) {
+											obj.add(entry.getKey(), jsonElement.getAsJsonArray());
+										} else {
+											obj.add(entry.getKey(), jsonElement.getAsJsonObject());
+										}
+									}
+								}
+							}
 							elementAdapter.write(out, obj);
 						}
 
@@ -167,7 +230,40 @@ public class UpdateCredentialsResponse {
 						public UpdateCredentialsResponse read(JsonReader in) throws IOException {
 							JsonElement jsonElement = elementAdapter.read(in);
 							validateJsonElement(jsonElement);
-							return thisAdapter.fromJsonTree(jsonElement);
+							JsonObject jsonObj = jsonElement.getAsJsonObject();
+							// store additional fields in the deserialized instance
+							UpdateCredentialsResponse instance = thisAdapter.fromJsonTree(jsonObj);
+							for (Map.Entry<String, JsonElement> entry : jsonObj.entrySet()) {
+								if (!openapiFields.contains(entry.getKey())) {
+									if (entry.getValue().isJsonPrimitive()) { // primitive type
+										if (entry.getValue().getAsJsonPrimitive().isString())
+											instance.putAdditionalProperty(
+													entry.getKey(), entry.getValue().getAsString());
+										else if (entry.getValue().getAsJsonPrimitive().isNumber())
+											instance.putAdditionalProperty(
+													entry.getKey(), entry.getValue().getAsNumber());
+										else if (entry.getValue().getAsJsonPrimitive().isBoolean())
+											instance.putAdditionalProperty(
+													entry.getKey(),
+													entry.getValue().getAsBoolean());
+										else
+											throw new IllegalArgumentException(
+													String.format(
+															"The field `%s` has unknown primitive type. Value: %s",
+															entry.getKey(),
+															entry.getValue().toString()));
+									} else if (entry.getValue().isJsonArray()) {
+										instance.putAdditionalProperty(
+												entry.getKey(),
+												gson.fromJson(entry.getValue(), List.class));
+									} else { // JSON object
+										instance.putAdditionalProperty(
+												entry.getKey(),
+												gson.fromJson(entry.getValue(), HashMap.class));
+									}
+								}
+							}
+							return instance;
 						}
 					}.nullSafe();
 		}